+++ src/mongo.service.ts
@@ -7,6 +7,7 @@ const DB_URL = "DB_URL";
 const DB_NAME = "DB_NAME";
 const DB_HOST = "DB_HOST";
 const DB_PORT = "DB_PORT";
+const DB_USER = "DB_USER";
 
 export interface MongoUrlParamsInterface {
   url: string
@@ -22,8 +23,18 @@ export interface MongoParamsInterface {
   options?: string
 }
 
+export type MongoStatusCallback = (configurations: {
+  host: string,
+  port: string,
+  database: string,
+  user?: string
+}) => void
+
 export class MongoConnector extends Service {
   connection: MongoClient
+
+  onConnected: MongoStatusCallback
+  onDisconnected: MongoStatusCallback
 
   constructor(params: MongoUrlParamsInterface | MongoParamsInterface) {
     let _params = new MongoConnectionParameters(params)
@@ -33,15 +44,32 @@ export class MongoConnector extends Service {
     map.set(DB_HOST, _params.host)
     map.set(DB_PORT, _params.port)
     map.set(DB_URL, _params.url)
+    map.set(DB_USER, _params.user)
     super(map)
   }
 
   async connect() {
     this.connection = new MongoClient(this.params.get(DB_URL))
+    if (this.onConnected != null) {
+      this.onConnected({
+        host: this.params.get(DB_HOST),
+        port: this.params.get(DB_PORT),
+        database: this.params.get(DB_NAME),
+        user: this.params.get(DB_USER)
+      })
+    }
   }
 
   async disconnect() {
     await this.connection.close()
+    if (this.onDisconnected != null) {
+      this.onDisconnected({
+        host: this.params.get(DB_HOST),
+        port: this.params.get(DB_PORT),
+        database: this.params.get(DB_NAME),
+        user: this.params.get(DB_USER)
+      })
+    }
   }
 
   override async beforeMount(context: Context) {